% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-total-records.R
\name{get_total_records}
\alias{get_total_records}
\title{Retrieve total number of publications from a request}
\usage{
get_total_records(query, database = "WOK")
}
\arguments{
\item{query}{a character of length 1. The query to send to the WOS lite API.}

\item{database}{a character of length 1. One among... See ... for further
information.}
}
\value{
The total number of publications (integer of length 1) matching the
query.
}
\description{
Retrieve total number of publications from a request
}
\examples{
query <- "TS=((salmo AND salar) AND conservation)"
get_total_records(query, database = "WOK")
}
