% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos-search.R
\name{wos_search}
\alias{wos_search}
\title{Retrieve total number of publications from a request}
\usage{
wos_search(query, database = "WOK")
}
\arguments{
\item{query}{a character of length 1. The query to send to the WOS lite API.}

\item{database}{a character of length 1. One among... See ... for further
information.}
}
\value{
The total number of publications (integer of length 1) matching the
query.
}
\description{
https://developer.clarivate.com/apis/woslite
}
\examples{
query <- "TS=((salmo AND salar) AND conservation)"
wos_search(query)
}
