% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos-search.R
\name{wos_search}
\alias{wos_search}
\title{Retrieve the number of records that match a given WOS query}
\usage{
wos_search(query, database = "WOS")
}
\arguments{
\item{query}{a \code{character} of length 1. The query to send to the WOS lite
API. Visit the WOS documentation at:
\url{https://images.webofknowledge.com/images/help/WOK/contents.html} on
how to write a WOS query.}

\item{database}{a \code{character} of length 1. One among \code{BCI} (BIOSIS Citation
Index), \code{BIOABS} (Biological Abstracts), \code{BIOSIS} (BIOSIS Previews),
\code{CCC} (Current Contents Connect), \code{DCI} (Data Citation Index), \code{DIIDW}
(Derwent Innovations Index), \code{MEDLINE} (Medline), \code{WOS} (Web of Science
Core Collection), \code{ZOOREC} (Zoological Record), and \code{WOK} (all databases).
Default is \code{WOS} (all databases).}
}
\value{
The total number of records (\code{integer} of length 1) that match the
query.
}
\description{
This function sends a query to the Web Of Science Lite API
(\url{https://developer.clarivate.com/apis/woslite}) and returns the total
number of records that match this query.

To learn how to write a WOS query, users can read the WOS documentation
available at:
\url{https://images.webofknowledge.com/images/help/WOK/contents.html}.
A list of WOS field tags is available at:
\url{https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html}.
}
\examples{
\dontrun{
## Search in TOPIC an exact expression ----
query <- "TS=\"salmo salar\""
wos_search(query)

## Search in TOPIC an exact expression and another term ----
query <- "TS=(\"salmo salar\" AND conservation)"
wos_search(query)

## Search for a specific year ----
query <- "TS=(\"salmo salar\" AND conservation) AND PY=2021"
wos_search(query)

## Search for a time span ----
query <- "TS=(\"salmo salar\" AND conservation) AND PY=2010-2021"
wos_search(query)

## Search for an author ----
query <- "AU=(\"Casajus N\")"
wos_search(query)
}
}
